fix: revert sqs dial optimizations
